Refactor agent architecture: Introduce AgentProvider to decouple chat and agent data management. Replace useAgentSettingsForm with useAgent for improved state handling. Update components to utilize new provider structure, enhancing flexibility for multiple chat instances and reducing global state pollution. Remove deprecated files and streamline data flow for better performance and maintainability.
