+++ apps/web/src/components/agent/breadcrumb-segment.tsx
@@ -1,42 +1,13 @@
-import { NotFoundError, useAgent } from "@deco/sdk";
-import { Suspense } from "react";
 import { AgentVisibility } from "../common/agent-visibility.tsx";
 import { AgentAvatar } from "../common/avatar/agent.tsx";
-import { ErrorBoundary } from "../../error-boundary.tsx";
+import { useAgent } from "./provider.tsx";
 
 interface Props {
-  agentId: string;
   variant?: "default" | "summary";
 }
 
-export function AgentBreadcrumbSegment({
-  agentId,
-  variant = "default",
-}: Props) {
-  return (
-    <ErrorBoundary
-      fallback={null}
-      shouldCatch={(e) => e instanceof NotFoundError}
-    >
-      <Suspense fallback={<AgentBreadcrumbSegment.Loading />}>
-        <AgentBreadcrumbSegment.UI agentId={agentId} variant={variant} />
-      </Suspense>
-    </ErrorBoundary>
-  );
-}
-
-AgentBreadcrumbSegment.Loading = () => {
-  return null;
-};
-
-AgentBreadcrumbSegment.UI = ({
-  agentId,
-  variant = "default",
-}: {
-  agentId: string;
-  variant: "default" | "summary";
-}) => {
-  const { data: agent } = useAgent(agentId);
+export function AgentBreadcrumbSegment({ variant = "default" }: Props) {
+  const { agent } = useAgent();
 
   return (
     <div className="flex items-center gap-2">
@@ -49,4 +20,4 @@ AgentBreadcrumbSegment.UI = ({
       <AgentVisibility agent={agent} />
     </div>
   );
-};
+}
