+++ apps/web/src/components/agent/index.ts
@@ -1,0 +1,13 @@
+export { AgentProvider, useAgent } from "./provider.tsx";
+export { default as AgentEdit } from "./edit.tsx";
+export { default as AgentChat } from "./chat.tsx";
+export { default as AgentPreview } from "./preview.tsx";
+export { default as AgentChats } from "./chats.tsx";
+export { default as AgentThread } from "./thread.tsx";
+export { default as AgentThreads } from "./threads.tsx";
+export { AgentBreadcrumbSegment } from "./breadcrumb-segment.tsx";
+export { ChatHeader as AgentChatHeader } from "./chat-header.tsx";
+export {
+  KnowledgeBaseFileList,
+  AddFileToKnowledgeButton,
+} from "./upload-knowledge-asset.tsx";
