+++ apps/web/src/components/agent/thread.tsx
@@ -1,5 +1,5 @@
 import { ChatMessages } from "../chat/chat-messages.tsx";
-import { ChatProvider } from "../chat/context.tsx";
+import { AgentProvider } from "./provider.tsx";
 import { ScrollArea } from "@deco/ui/components/scroll-area.tsx";
 import { ChatHeader } from "./chat-header.tsx";
 
@@ -14,14 +14,14 @@ function ThreadView({ agentId, threadId }: Props) {
   }
 
   return (
-    <ChatProvider agentId={agentId} threadId={threadId}>
+    <AgentProvider agentId={agentId} threadId={threadId}>
       <div className="flex items-center justify-between p-4">
         <ChatHeader />
       </div>
       <ScrollArea className="h-full w-full p-6 text-foreground">
         <ChatMessages />
       </ScrollArea>
-    </ChatProvider>
+    </AgentProvider>
   );
 }
