+++ apps/web/src/components/audit/detail.tsx
@@ -2,7 +2,7 @@ import { useThread } from "@deco/sdk";
 import { ScrollArea } from "@deco/ui/components/scroll-area.tsx";
 import { useParams } from "react-router";
 import { ChatMessages } from "../chat/chat-messages.tsx";
-import { ChatProvider } from "../chat/context.tsx";
+import { AgentProvider } from "../agent/provider.tsx";
 import type { Tab } from "../dock/index.tsx";
 import { DefaultBreadcrumb, PageLayout } from "../layout.tsx";
 
@@ -38,7 +38,7 @@ function Page() {
   const { data: { title } = {} } = useThread(id);
 
   return (
-    <ChatProvider agentId={thread?.metadata?.agentId ?? id} threadId={id}>
+    <AgentProvider agentId={thread?.metadata?.agentId ?? id} threadId={id}>
       <PageLayout
         hideViewsButton
         tabs={TABS}
@@ -51,7 +51,7 @@ function Page() {
           />
         }
       />
-    </ChatProvider>
+    </AgentProvider>
   );
 }
