+++ apps/web/src/components/chat/empty-state.tsx
@@ -1,14 +1,14 @@
-import { NotFoundError, useAgent, WELL_KNOWN_AGENT_IDS } from "@deco/sdk";
+import { NotFoundError, WELL_KNOWN_AGENT_IDS } from "@deco/sdk";
 import { Button } from "@deco/ui/components/button.tsx";
 import { Icon } from "@deco/ui/components/icon.tsx";
 import { Suspense } from "react";
 import { ErrorBoundary } from "../../error-boundary.tsx";
 import { useFocusChat } from "../agents/hooks.ts";
 import { AgentAvatar } from "../common/avatar/agent.tsx";
-import { useChatContext } from "./context.tsx";
+import { useAgent } from "../agent/provider.tsx";
 
 export function EmptyState() {
-  const { agentId } = useChatContext();
+  const { agentId } = useAgent();
 
   if (agentId === WELL_KNOWN_AGENT_IDS.teamAgent) {
     return (
@@ -56,8 +56,7 @@ EmptyState.Skeleton = () => {
 };
 
 EmptyState.UI = () => {
-  const { agentId, uiOptions } = useChatContext();
-  const { data: agent } = useAgent(agentId);
+  const { agentId, agent, uiOptions } = useAgent();
   const editAgent = useFocusChat();
 
   return (
