+++ apps/web/src/components/chat/extensions/tool-suggestion.ts
@@ -75,7 +75,7 @@ export const suggestion: (
             showOnCreate: true,
             interactive: true,
             trigger: "manual",
-            placement: "bottom-start",
+            placement: "top-start",
             maxWidth: 400,
           });
         },
