+++ apps/web/src/components/decopilot/index.tsx
@@ -1,8 +1,7 @@
 import { WELL_KNOWN_AGENTS } from "@deco/sdk";
 import { DockviewApi } from "dockview-react";
 import { MainChat } from "../agent/chat.tsx";
-import { ChatProvider } from "../chat/context.tsx";
-import { AgentSettingsFormProvider } from "../agent/agent-settings-form-provider.tsx";
+import { AgentProvider } from "../agent/provider.tsx";
 
 export const NO_DROP_TARGET = "no-drop-target";
 
@@ -36,7 +35,7 @@ export const toggleDecopilotTab = (api: DockviewApi) => {
 
 export function DecopilotChat() {
   return (
-    <ChatProvider
+    <AgentProvider
       agentId={WELL_KNOWN_AGENTS.decopilotAgent.id}
       threadId="1"
       uiOptions={{
@@ -47,10 +46,8 @@ export function DecopilotChat() {
         showEditAgent: false,
       }}
     >
-      <AgentSettingsFormProvider agentId={WELL_KNOWN_AGENTS.decopilotAgent.id}>
-        <MainChat />
-      </AgentSettingsFormProvider>
-    </ChatProvider>
+      <MainChat />
+    </AgentProvider>
   );
 }
 DecopilotChat.displayName = "DefaultChat";
