+++ apps/web/src/components/settings/channels.tsx
@@ -34,7 +34,7 @@ import { toast } from "@deco/ui/components/sonner.tsx";
 import { Spinner } from "@deco/ui/components/spinner.tsx";
 import { cn } from "@deco/ui/lib/utils.ts";
 import { useMemo, useState } from "react";
-import { useAgentSettingsForm } from "../agent/agent-settings-form-provider.tsx";
+import { useAgent } from "../agent/provider.tsx";
 import { IntegrationIcon } from "../integrations/common.tsx";
 import { InstalledConnections } from "../integrations/installed-connections.tsx";
 import { Dialog } from "@deco/ui/components/dialog.tsx";
@@ -95,7 +95,7 @@ export function Channels({ className }: ChannelsProps) {
     channelId: string;
     channelName: string;
   } | null>(null);
-  const { agent } = useAgentSettingsForm();
+  const { agent } = useAgent();
   const { mutate: createChannel, isPending: isCreating } = useCreateChannel();
   const joinChannelMutation = useJoinChannel();
   const leaveChannelMutation = useLeaveChannel();
