+++ apps/web/src/hooks/use-tools.ts
@@ -1,8 +1,8 @@
-import { useAgent } from "@deco/sdk";
+import { useAgentData } from "@deco/sdk";
 import { useMemo } from "react";
 
 export const useTools = (agentId?: string) => {
-  const { data: agent } = agentId ? useAgent(agentId) : { data: null };
+  const { data: agent } = agentId ? useAgentData(agentId) : { data: null };
 
   return useMemo(() => agent?.tools_set ?? {}, [agent]);
 };
